Patch to ArrayOutOfBoundsIndex exception inside PolygonModel method:
Changes made to the PolygonModel class which now supports 2D arrays with
different lengths of columns in each row. 
Arrays xp and yp, which are passed into the DrawPolygon method are now
initiated and resized according to the length of the row that the array
is currently on.

Errors given PrePatch: Applet Crashed due to ArrayOutOfBounds exception.

(Out of class Debugging).